pyproject.toml
[tool.poetry]
name = "flask_psyco_template"
version = "0.1.0"
description = ""
authors = ["Alexander Lavrov <lavrov@selectel.ru>"]
license = "ISC"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1"
flask_sqlalchemy = "^2.4"
flask-migrate = "^2.5"
flask-potion = "^0.16.0"
psycopg2-binary = "^2.8"
flask-dramatiq = "^0.4.1"
redis = "^3.3"
pre-commit = "^1.18"
pip = "^19.2"
passlib = "^1.7"
flask_jwt = "^0.3.2"
flask-shell-ipython = "^0.4.1"
python-slugify = "^3.0"

[tool.poetry.dev-dependencies]
bumpversion = "^0.5.3"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )
)
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
